final check

+++ server_app/back/src/main/java/com/bigp/back/controller/ConfigApiController.java
@@ -27,10 +27,12 @@ import com.bigp.back.entity.AdminInfo;
 import com.bigp.back.entity.BabyInfo;
 import com.bigp.back.entity.ConfigInfo;
 import com.bigp.back.entity.UserInfo;
+import com.bigp.back.service.AESService;
 import com.bigp.back.service.AdminInfoService;
 import com.bigp.back.service.BabyInfoService;
 import com.bigp.back.service.ConfigInfoService;
 import com.bigp.back.service.UserInfoService;
+import com.bigp.back.utils.CheckUtils;
 import com.bigp.back.utils.JwtTokenProvider;
 
 import lombok.RequiredArgsConstructor;
@@ -46,20 +48,22 @@ public class ConfigApiController {
     private final AdminInfoService adminInfoService;
     private final BabyInfoService babyService;
     private final ConfigInfoService configService;
+    private final CheckUtils checkUtils;
+    private final AESService aesService;
 
     @GetMapping("/getSettingInfo")
     public ResponseEntity<?> getSettingInfo(@CookieValue(name="accessToken", required=true) String accessToken) {
 
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)) {
+                UserInfo user = userService.getEntity("accessToken", accessToken);
 
                 if (user != null) {
                     BabyInfo baby = user.getBabyInfo();
                     ConfigInfo config = user.getConfigInfo();
 
                     if (baby != null && config != null) {
-                        return ResponseEntity.ok(new GetSettingInfoApiDto.SuccessResponse(true, config.isAlarm(), baby.getBabyname(), baby.getBabybirth(), config.getDataeliminateduration()));
+                        return ResponseEntity.ok(new GetSettingInfoApiDto.SuccessResponse(true, config.isAlarm(), aesService.decryptInfo(baby.getBabyname()), aesService.decryptInfo(baby.getBabybirth()), config.getDataeliminateduration()));
                     }
                 }
             }
@@ -80,8 +84,9 @@ public class ConfigApiController {
         int dataEliminateDuration = request.getDataEliminateDuration();
 
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)
+                && !checkUtils.checkQuery(babyName) && !checkUtils.checkQuery(babyBirth)) {
+                UserDTO.UserInfo user = userService.getUserInfo("accessToken", accessToken);
                 if (user != null) {
                     UserDTO.BabyInfo baby = new UserDTO.BabyInfo();
                     UserDTO.ConfigInfo config = new UserDTO.ConfigInfo();
@@ -109,14 +114,13 @@ public class ConfigApiController {
         String accessToken = request.getAccessToken();
 
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)) {
+                UserDTO.UserInfo user = userService.getUserInfo("accessToken", accessToken);
                 AdminInfo admin = adminInfoService.getAdminInfo(accessToken);
                 if (user != null) {
-                    UserDTO.UserInfo userDto = new UserDTO.UserInfo();
 
-                    userDto.setAccessToken(accessToken + ".logout");
-                    boolean isUpdate = userService.updateUser("accessToken", accessToken, userDto);
+                    user.setAccessToken(accessToken + ".logout");
+                    boolean isUpdate = userService.updateUser("accessToken", accessToken, user);
 
                     if (isUpdate)
                         return ResponseEntity.ok(new LogoutApiDto.SuccessResponse(true));
@@ -143,8 +147,8 @@ public class ConfigApiController {
         String accessToken = request.getAccessToken();
 
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)) {
+                UserDTO.UserInfo user = userService.getUserInfo("accessToken", accessToken);
 
                 if (user != null) {
                     if (userService.deleteUser(accessToken))
@@ -163,11 +167,11 @@ public class ConfigApiController {
     public ResponseEntity<?> getProfileImage(@CookieValue(name="accessToken", required=true) String accessToken) {
 
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)) {
+                UserDTO.UserInfo user = userService.getUserInfo("accessToken", accessToken);
 
                 if (user != null) {
-                    String profileImage = Base64.getEncoder().encodeToString(user.getProfileImage());
+                    String profileImage = user.getProfileImage();
                     return ResponseEntity.ok(new GetProfileImageApiDto.SuccessResponse(true, profileImage));
                 }
             }
@@ -180,7 +184,7 @@ public class ConfigApiController {
     }
 
     @PostMapping("/setProfileImage")
-    public ResponseEntity<?> postMethodName(@RequestBody SetProfileImageApiDto.RequestBody request) {
+    public ResponseEntity<?> setProfileImage(@RequestBody SetProfileImageApiDto.RequestBody request) {
         String accessToken = request.getAccessToken();
         String profile = request.getProfileImage();
         int isSuccess = userService.setProfileImage("accessToken", accessToken, profile);
@@ -197,8 +201,8 @@ public class ConfigApiController {
     @GetMapping("/getPersonalInfo")
     public ResponseEntity<?> getPersonalInfo(@CookieValue(name="accessToken", required=true) String accessToken) {
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)) {
+                UserDTO.UserInfo user = userService.getUserInfo("accessToken", accessToken);
 
                 if (user != null)
                     return ResponseEntity.ok(new GetPersonalInfo.SuccessResponse(true, user.getUsername(), user.getEmail()));
@@ -214,8 +218,8 @@ public class ConfigApiController {
     @GetMapping("/getAliasName")
     public ResponseEntity<?> getAliasName(@CookieValue(name="accessToken", required=true) String accessToken) {
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
-                UserInfo user = userService.getUserInfo("accessToken", accessToken);
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)) {
+                UserDTO.UserInfo user = userService.getUserInfo("accessToken", accessToken);
 
                 if (user != null)
                     return ResponseEntity.ok(new GetAliasNameApiDto.SuccessResponse(true, user.getAliasname()));
@@ -234,7 +238,8 @@ public class ConfigApiController {
         String name = request.getName();
         
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
+            if (jwtTokenProvider.isExpired(accessToken) || !checkUtils.checkQuery(accessToken)
+                && !checkUtils.checkQuery(name)) {
                 UserDTO.UserInfo userDto = new UserDTO.UserInfo();
 
                 userDto.setAliasname(name);
@@ -260,7 +265,8 @@ public class ConfigApiController {
         String password = request.getPassword();
         
         try {
-            if (jwtTokenProvider.isExpired(accessToken)) {
+            if (jwtTokenProvider.isExpired(accessToken) && !checkUtils.checkQuery(accessToken)
+                && !checkUtils.checkQuery(password)) {
                 UserDTO.UserInfo user = new UserDTO.UserInfo();
 
                 user.setPassword(password);
